server/requirements.txt
beautifulsoup4==4.10.0
boto3==1.18.49
botocore==1.21.49
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.6
colorama==0.4.4
idna==3.2
jmespath==0.10.0
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
